requirements.txt
marker-pdf[full]
pylatex
pillow
transformers
torch